Rydder opp i ubrukt registrering-kafka kode
