+++ pom.xml
@@ -17,7 +17,6 @@
     <properties>
         <java.version>21</java.version>
         <common.version>3.2023.10.23_12.41-bafec3836d28</common.version>
-        <veilarbregistrering.skjema.version>c22beae</veilarbregistrering.skjema.version>
         <jaxb.version>2.3.1</jaxb.version>
         <testcontainers.version>1.20.0</testcontainers.version>
 
@@ -201,12 +200,6 @@
                     <artifactId>swagger-annotations</artifactId>
                 </exclusion>
             </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.navikt</groupId>
-            <artifactId>veilarbregistrering-skjema</artifactId>
-            <version>${veilarbregistrering.skjema.version}</version>
         </dependency>
         <dependency>
             <groupId>com.github.navikt</groupId>
