--- src/main/java/no/nav/pto/veilarbportefolje/arbeidssoeker/v1/registrering/endring/EndringIArbeidssokerRegistreringDTO.kt
@@ -1,9 +1,0 @@
-package no.nav.pto.veilarbportefolje.arbeidssoeker.v1.registrering.endring
-
-import java.time.LocalDate
-
-data class EndringIArbeidssokerRegistreringDTO(
-    val aktorId: String ? = null,
-    val brukersSituasjon: String ? = null,
-    val brukersSituasjonSistEndret: LocalDate? = null
-)
